Update mmd-to-iso.xsl
This change add a string replacement to fix an issue we have with `pycsw` - we cannot have commas in the `OGC:WMS` description - (the next major version of pycsw will fix this) - the underlying `pycsw` RDBMS model stores link properties (name, protocol, description, url) as a CSV, so the comma in the `OGC:WMS` description trips up the serializer. It’s a known issue, this PR allow us to work around it. -- we may want to check all the others link properties as well.